Split genome class from network
