+++ src/players/net.py
@@ -1,14 +1,7 @@
 from game.action import DIRECTIONS
+from genetics.genome import Genome
 import numpy as np
 from players.base import Player
-
-
-NUM_HIDDEN_LAYERS = 2
-assert NUM_HIDDEN_LAYERS > 0
-HIDDEN_LAYER_SIZE = 256
-CHROMOSOME_SIZE = (16 * HIDDEN_LAYER_SIZE +  # Input layer weights
-                   (NUM_HIDDEN_LAYERS - 1) * HIDDEN_LAYER_SIZE**2 +  # Hidden layer weights
-                   HIDDEN_LAYER_SIZE * 4)  # Output layer weights
 
 
 class NetworkPlayer(Player):
@@ -23,12 +16,12 @@ class NetworkPlayer(Player):
     ----------
     generation : int
         Which generation the network belongs to.
-    chromosome : ndarray
+    genome : ndarray
         A 340x1 numpy array containing the weights for the matrices. (16 inputs; 1 hidden layer of size 16, plus bias.
         17x16 matrix then 17x4 = 340 elements)
     """
 
-    def __init__(self, gen=0, mom=None, dad=None, chromosome=None):
+    def __init__(self, gen=0, mom=None, dad=None, genome=None):
         """Builds the network from a chromosome if given, or two parents, falling back to random generation if neither.
 
         Parameters
@@ -39,27 +32,17 @@ class NetworkPlayer(Player):
             A net from which the chromosome will be sampled.
         dad : Optional[Net]
             The other net from which the chromosome will be sampled.
-        chromosome : Optional[ndarray]
+        genome : Optional[ndarray]
             A 340x1 numpy array containing the network weights.
         """
         super().__init__()
         self.generation = gen
-        if chromosome is not None:
-            self.chromosome = chromosome
+        if genome is not None:
+            self.genome = genome
         elif None not in [mom, dad]:
-            self.chromosome = np.array([
-                    mom.chromosome[i] if np.random.random() > 0.5
-                    else dad.chromosome[i]
-                    for i in range(len(mom.chromosome))
-                    ])
-            self._mutate()
+            self.genome = Genome(mom.genome, dad.genome)
         else:
-            self.chromosome = 2 * np.random.randint(0, 2, CHROMOSOME_SIZE) - 1
-
-    def _mutate(self):
-        """Add random mutations to 2% of net's chromosome."""
-        mutation = np.array([-1 if np.random.random() < 0.01 else 1 for _ in range(len(self.chromosome))])
-        self.chromosome *= mutation
+            self.genome = Genome()
 
     def _choose_action(self, game):
         """Evaluate the position using the network and choose the best legal move it determines.
@@ -93,10 +76,7 @@ class NetworkPlayer(Player):
         ndarray
             The four direction actions sorted in the order of the network's evaluation.
         """
-        w_xh = self.chromosome[:16 * HIDDEN_LAYER_SIZE].reshape((16, HIDDEN_LAYER_SIZE))
-        w_hh = self.chromosome[16 * HIDDEN_LAYER_SIZE:-HIDDEN_LAYER_SIZE * 4]\
-            .reshape((NUM_HIDDEN_LAYERS - 1, HIDDEN_LAYER_SIZE, HIDDEN_LAYER_SIZE))
-        w_hy = self.chromosome[-HIDDEN_LAYER_SIZE * 4:].reshape((HIDDEN_LAYER_SIZE, 4))
+        w_xh, w_hh, w_hy = self.genome.get_weight_matrices()
 
         x = 3 * (board.reshape(16) / 7 - 1)  # Max tile log-value in 2048 is 14. Normalize to [-3, 3].
         h = np.sign(x @ w_xh)
